refactor(places-service): rename everything from places to places service
BREAKING CHANGES: hook, props, returns, example files were renamed from usePlaces to usePlacesService
