+++ examples/places-service/components/map-canvas/map-canvas.module.css
