+++ examples/places-service/index.html
