+++ package.json
@@ -12,8 +12,8 @@
     "start:example": "./start-example.sh",
     "start:map-example": "EXAMPLE=map run-p start:library start:example",
     "start:map-with-markers-example": "EXAMPLE=map-with-markers run-p start:library start:example",
-    "start:geocoding-example": "EXAMPLE=geocoding run-p start:library start:example",
-    "start:places-example": "EXAMPLE=places run-p start:library start:example",
+    "start:geocoding--example": "EXAMPLE=geocoding run-p start:library start:example",
+    "start:places-service-example": "EXAMPLE=places-service run-p start:library start:example",
     "start:places-autocomplete-example": "EXAMPLE=places-autocomplete run-p start:library start:example",
     "start:directions-service-example": "EXAMPLE=directions-service run-p start:library start:example",
     "start:distance-matrix-example": "EXAMPLE=distance-matrix run-p start:library start:example",
